refactor: separate express and socket.io
