+++ server/src/express/routes/protected.ts
