+++ server/src/socket.io/namespaces/admin/events/user.ts
@@ -1,6 +1,6 @@
-import { updateUser } from "../../../services/user.service";
-import { TUserIO, TUserSocket } from "../../../types/socket.io/admin";
-import { IUser } from "../../../types/user";
+import { updateUser } from "../../../../services/user.service";
+import { TUserIO, TUserSocket } from "../../../@types/admin";
+import { IUser } from "../../../../types/user";
 
 function registerUserHandlers(_io: TUserIO, socket: TUserSocket) {
   async function onStatusUpdate(status: IUser["status"]) {
