+++ server/src/socket.io/namespaces/admin/index.ts
@@ -1,11 +1,13 @@
 import { Server, Socket } from "socket.io";
-import { authentication } from "../../middlewares/auth";
+import { authenticate, authorize } from "../../middlewares/auth";
 import registerUserHandlers from "./events/user";
 
 function registerAdminNamespace(io: Server) {
   const admin = io.of("/admin");
 
-  admin.use(authentication.socketIO);
+  admin.use(authenticate);
+
+  admin.use(authorize);
 
   function onConnection(socket: Socket) {
     console.log("[socket.io]: Admin connected", socket.request.user?.id);
